requirements/base.txt
# https://docs.djangoproject.com/en/stable/
Django==4.0.1

# https://django-allauth.readthedocs.io/en/stable/
django-allauth==0.47.0

# https://django-compressor.readthedocs.io/en/stable/
django-compressor==3.1

# https://django-crispy-forms.readthedocs.io/en/stable
django-crispy-forms==1.13.0

# https://django-environ.readthedocs.io/en/stable/
django-environ==0.8.1

# https://github.com/neo4j-contrib/django-neomodel
django-neomodel==0.0.7

# https://django-tables2.readthedocs.io/en/stable/
django-tables2

# https://dateutil.readthedocs.io/en/stable/
python-dateutil

# https://github.com/skorokithakis/shortuuid
shortuuid==1.0.8
